finished ticks to rotations conversion
